requirements.txt
Twisted==24.7.0
Klein==23.12.0
txpostgres==1.7.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
Jinja2==3.1.4
python-dotenv==1.0.1
pytest==8.3.2
pytest-twisted==1.14.3
treq==23.11.0
twisted-iocpsupport==1.0.4
